requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
anyio==3.6.2
attrs==22.2.0
autopep8==2.0.2
boto3==1.26.89
botocore==1.29.89
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.94.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
loguru==0.6.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.6
python-dateutil==2.8.2
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.6
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.0.post1
typing_extensions==4.5.0
urllib3==1.26.15
win32-setctime==1.1.0
